Add busybox-statefulset application
Like the busybox deployment using 2 replicas and statefulset.

The application uses the same labels and names as the busybox
application to make it easy to test both applications. The only
difference is the namespace: "busybox-sample-statefulset" instead of
"busybox-sample". This way we can run both application in the same
cluster concurrently.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>