Describe the sample applications
List the application for OpenShift and how to use them with the
OpenShift Console (deploy, enable-dr) or command line (disable DR).

List the applications for Kubernetes and how to use them with kubectl
commands.

We show example commands only for the busybox application; for other
applications you need to use the relevant kustomization and namespace.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>