Move channel into the subscription
Use the same structure as ocm samplesp[1], keeping the channel in the
subscription kustomization under the application namespace.

This is better for users, removing one step from the deployment steps.
Now it is not possible to forget to deploy the channel.

This is better for developers, since now we can test multiple
applications concurrently, without dealing with setup and cleanup of
shared resources, since application do not share the channel.

[1] https://github.com/open-cluster-management-io/multicloud-operators-subscription/tree/main/examples/github-channel

Signed-off-by: Nir Soffer <nsoffer@redhat.com>