Use distinct labels for pvcs and kube objects
Previously we use the same label (appname: busybox) for the deployment,
pods, and pvc. When using OCM managed applications this works fine -
ramen backup and restore the PVC by the label, and for the rest of the
resources, the label is handy for looking up the application. But when
using kube object protection, velero will backup all resources including
generated resources such as pods and the pvc backed up by ramen. Then
velero restore all the resources and may fail to restore if ramen
already modified the restored pvc.

With the new labels, one can use distinct labels:

- kube object selector: "protected-by" In ["ramen"]
- pvc selector: "replicated-by" In ["ramen"]

With this velero will backup only the deployment, which will create the
pods, which will use the ramen restored pvc - same flow as OCM managed
applications.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>