Add discovered deployment-rbd dr resources
The ManagedClusterBinding belongs to the ramen-ops namespace and
not to the application, so we keep it in the ramen-ops directory.

The rest of the resources are in the deployment-rbd directory. I'm not
sure how easy it will be to share a base kustomization with other
workloads so lets start with something simple.

The README was updated to describe a complete DR flow for discovered
application using the new resources.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>