Add Windows server 2022 demo VM
We want to use this for a demo, showing that we can protect Windows VMs.
The kubevirt environment memory, cpus and disk are increased to allow
testing the windows VM in drenv environment.

The VM was tested with container disk created manually. Users need to
use they own container disk since the Windows image is not
distributable.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>