Add Linux demo vm
Add Fedora demo VM using Fedora 40 image created with virt-builder.
The demo VM includes the visitor demo application.

The demo VM includes overlays for Kubernetes and OpenShift.

The VM does not use a public key since I could not get cloud-init to
work with the Fedora image. We can try to fix this later or use Fedora
cloud image which comes with configured cloud-init, and use cloud-init
to install and setup the visitor application.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>