Add configmap
Using the configmap we can inject cluster info into the VM. To use this
we need to create the VM namespace on all clusters and create the
example configmap in the namespace. We will use this to show the actual
cluster name instead of the host from the http request.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>